Extract CSP configs
